Horde: Refactored system for creating bundles of tree objects.
- Tree can now be modified without having to have entire working set in memory.
- Multiple nodes can be compressed/decompressed together, and an optional IMemoryCache allows caching decompressed blocks.
- Added support for specifying custom compact-binary serializers on individual properties.
- Added test suite to test bundling functionality.
- Fixed issue where VarInts of negative values would not serialize correctly.
#preflight 621952f68849a7424fa36567

[CL 19158249 by Ben Marsh in ue5-main branch]

+++ Engine/Source/Programs/Shared/EpicGames.Serialization/VarInt.cs
@@ -60,7 +60,7 @@ namespace EpicGames.Serialization
 		/// <returns></returns>
 		public static int Measure(int Value)
 		{
-			return Measure((uint)Value);
+			return Measure((ulong)(long)Value);
 		}
 
 		/// <summary>
